requirements.txt
numpy==1.24.3
matplotlib==3.7.2
langchain>=0.1.0  # Updated version
langchain-community>=0.1.0  # Updated version
scipy==1.11.3
websocket-client==1.6.4
streamlit==1.47.1
ollama==0.9.6

